requirements.txt
google-genai
groq
python-dotenv
sentence-transformers
faiss-cpu
PyPDF2
numpy